Global Moran's I statistic

Calculate the global Moran's I statistic for model residuals. `ww_global_moran_i()` returns the statistic itself, while `ww_global_moran_pvalue()` returns the associated p value. `ww_global_moran()` returns both.

ww_global_moran_i(data, ...)

ww_global_moran_i_vec(truth, estimate, nb, wt, na_rm = TRUE, ...)

ww_global_moran_pvalue(data, ...)

ww_global_moran_pvalue_vec(
  truth,
  estimate,
  nb,
  wt,
  alternative,
  randomization,
  na_rm = TRUE,
  ...
)

ww_global_moran(
  data,
  truth,
  estimate,
  nb,
  wt,
  alternative = "greater",
  randomization = TRUE,
  na_rm = TRUE,
  ...
)

Arguments

data

A data.frame containing the columns specified by the truth and estimate arguments.

...

Not currently used.

truth

The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate

The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

nb

a neighbor list object for example as created by st_contiguity().

wt

a weights list as created by st_weights().

na_rm

A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A tibble with columns .metric, .estimator, and .estimate. For grouped data frames, the number of rows returned will be the same as the number of groups times the number of metrics. For ww_global_moran_i_vec(), a single numeric value (or NA).