Local Geary's C statistic

Calculate the local Geary's C statistic for model residuals. ww_local_geary_c() returns the statistic itself, while ww_local_geary_pvalue() returns the associated p value. These functions are meant to help assess model predictions, for instance by identifying clusters of higher residuals than expected. For statistical testing and inference applications, use spdep::localC_perm() instead.

Usage

ww_local_geary_c(data, ...)

ww_local_geary_c_vec(truth, estimate, wt, na_action = na.fail, ...)

ww_local_geary_pvalue(data, ...)

ww_local_geary_pvalue_vec(truth, estimate, wt = NULL, na_action = na.fail, ...)

Arguments

data

A data.frame containing the columns specified by the truth and estimate arguments.

...

Additional arguments passed to spdep::localC() (for ww_local_geary_c()) or spdep::localC_perm() (for ww_local_geary_pvalue()).

truth

The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate

The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

wt

A listw object, for instance as created with ww_build_weights(). For data.frame input, may also be a function that takes data and returns a listw object.

na_action

A function which indicates what should happen when the data contain NAs. The default is na.fail; you may wish to set it to na.omit or any of the functions from the "zoo" package. This function ignores the value of options("na.action") in order to make cross-computer (and cross-session) results more stable. This argument must be of length 1; you cannot pass multiple functions to na_action.

Value

A tibble with columns .metric, .estimator, and .estimate and nrow(data)

rows of values. For _vec() functions, a numeric vector of length(truth) (or NA).

Details

These functions can be used for geographic or projected coordinate reference systems and expect 2D data.

References

Anselin, L. 1995. Local indicators of spatial association, Geographical Analysis, 27, pp 93–115. doi: 10.1111/j.1538-4632.1995.tb00338.x.

Anselin, L. 2019. A Local Indicator of Multivariate Spatial Association: Extending Geary's C. Geographical Analysis, 51, pp 133-150. doi: 10.1111/gean.12164

See also

Examples

guerry_model <- guerry
guerry_lm <- lm(Crm_prs ~ Litercy, guerry_model)
guerry_model$predictions <- predict(guerry_lm, guerry_model)

ww_local_geary_c(guerry_model, Crm_prs, predictions)
#> Warning: Returning more (or less) than 1 row per `summarise()` group was deprecated in
#> dplyr 1.1.0.
#> ℹ Please use `reframe()` instead.
#> ℹ When switching from `summarise()` to `reframe()`, remember that `reframe()`
#>   always returns an ungrouped data frame and adjust accordingly.
#> ℹ The deprecated feature was likely used in the yardstick package.
#>   Please report the issue at <https://github.com/tidymodels/yardstick/issues>.
#> # A tibble: 85 × 3
#>    .metric       .estimator .estimate
#>    <chr>         <chr>          <dbl>
#>  1 local_geary_c standard       0.981
#>  2 local_geary_c standard       0.836
#>  3 local_geary_c standard       0.707
#>  4 local_geary_c standard       0.108
#>  5 local_geary_c standard       0.264
#>  6 local_geary_c standard       1.36 
#>  7 local_geary_c standard       3.64 
#>  8 local_geary_c standard       1.57 
#>  9 local_geary_c standard       0.867
#> 10 local_geary_c standard       0.737
#> # … with 75 more rows
ww_local_geary_pvalue(guerry_model, Crm_prs, predictions)
#> # A tibble: 85 × 3
#>    .metric            .estimator .estimate
#>    <chr>              <chr>          <dbl>
#>  1 local_geary_pvalue standard       0.243
#>  2 local_geary_pvalue standard       0.226
#>  3 local_geary_pvalue standard       0.167
#>  4 local_geary_pvalue standard       0.139
#>  5 local_geary_pvalue standard       0.310
#>  6 local_geary_pvalue standard       0.133
#>  7 local_geary_pvalue standard       0.369
#>  8 local_geary_pvalue standard       0.194
#>  9 local_geary_pvalue standard       0.739
#> 10 local_geary_pvalue standard       0.419
#> # … with 75 more rows

wt <- ww_build_weights(guerry_model)

ww_local_geary_c_vec(
  guerry_model$Crm_prs,
  guerry_model$predictions,
  wt = wt
)
#>  [1] 0.981119438 0.836402177 0.707464373 0.108332465 0.264075824 1.361485477
#>  [7] 3.641239412 1.571824022 0.867252524 0.737094462 0.573376555 0.001605731
#> [13] 1.891988440 1.152840284 1.029320931 0.297642850 1.219953394 1.934113868
#> [19] 1.632566652 0.441916658 5.202733790 0.921953310 3.084515822 0.237218594
#> [25] 1.346684045 1.051652204 0.419414691 0.217280214 0.794409207 0.243971372
#> [31] 0.376678958 0.139152907 0.711305633 3.096840680 1.974463944 0.922230710
#> [37] 1.032031759 0.339464386 0.933794842 1.910440700 0.937597672 0.625628647
#> [43] 0.376707677 2.692250283 1.288784962 0.798443065 1.671895951 1.310183326
#> [49] 2.347513577 0.845204889 0.302940809 2.291804447 0.881999216 0.412051312
#> [55] 2.006031605 0.561239582 0.375776092 1.853716391 1.191472387 1.146802970
#> [61] 1.857618679 0.149044974 0.614228825 0.755373475 1.287962784 1.447534518
#> [67] 1.236607966 0.962394651 0.338400653 1.914478855 0.641340157 2.146993342
#> [73] 0.703881855 1.417638272 0.692636715 1.765618175 0.246058853 0.700262130
#> [79] 0.002876896 0.057575267 0.420878038 2.025012395 2.525093274 1.053335832
#> [85] 1.030009749
ww_local_geary_pvalue_vec(
  guerry_model$Crm_prs,
  guerry_model$predictions,
  wt = wt
)
#>  [1] 0.21806108 0.24623925 0.16643898 0.12466159 0.31848226 0.18048630
#>  [7] 0.46491080 0.18001567 0.79431499 0.42832156 0.03135908 0.16180581
#> [13] 0.26788150 0.89849244 0.57100001 0.21322066 0.64900941 0.52551969
#> [19] 0.07595116 0.17416114 0.57250357 0.77915907 0.59235100 0.13268454
#> [25] 0.88214171 0.95829755 0.20784564 0.07871680 0.60549546 0.13250885
#> [31] 0.18918685 0.14604256 0.45149101 0.63057656 0.07800566 0.86487561
#> [37] 0.34819967 0.27078087 0.87972322 0.87175355 0.65402676 0.53778895
#> [43] 0.19911582 0.23185270 0.81970417 0.05397182 0.26821732 0.16470964
#> [49] 0.17346332 0.23208333 0.05123291 0.74807819 0.42495951 0.24808302
#> [55] 0.95487698 0.17724240 0.21396133 0.60843588 0.24101719 0.94914151
#> [61] 0.27521252 0.28662809 0.31732083 0.19387877 0.55052932 0.37391106
#> [67] 0.70539573 0.79663838 0.02410305 0.07253076 0.50146578 0.13399955
#> [73] 0.36297254 0.54971214 0.53783865 0.09527843 0.11444653 0.31824812
#> [79] 0.16543064 0.05574109 0.32976950 0.53793451 0.08172541 0.97845702
#> [85] 0.99338310