Local Getis-Ord G and G* statistic

Calculate the local Getis-Ord G and G* statistic for model residuals. ww_local_getis_ord_g() returns the statistic itself, while ww_local_getis_ord_pvalue() returns the associated p value. These functions are meant to help assess model predictions, for instance by identifying clusters of higher residuals than expected. For statistical testing and inference applications, use spdep::localG_perm() instead.

Usage

ww_local_getis_ord_g(data, ...)

ww_local_getis_ord_g_vec(truth, estimate, wt, na_action = na.fail, ...)

ww_local_getis_ord_g_pvalue(data, ...)

ww_local_getis_ord_g_pvalue_vec(truth, estimate, wt, na_action = na.fail, ...)

Arguments

data

A data.frame containing the columns specified by the truth and estimate arguments.

...

Additional arguments passed to spdep::localG() (for ww_local_getis_ord_g()) or spdep::localG_perm() (for ww_local_getis_ord_pvalue()).

truth

The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate

The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

wt

A listw object, for instance as created with ww_build_weights(). For data.frame input, may also be a function that takes data and returns a listw object.

na_action

A function which indicates what should happen when the data contain NAs. The default is na.fail; you may wish to set it to na.omit or any of the functions from the "zoo" package. This function ignores the value of options("na.action") in order to make cross-computer (and cross-session) results more stable. This argument must be of length 1; you cannot pass multiple functions to na_action.

Value

A tibble with columns .metric, .estimator, and .estimate and nrow(data)

rows of values. For _vec() functions, a numeric vector of length(truth) (or NA).

Details

These functions can be used for geographic or projected coordinate reference systems and expect 2D data.

References

Ord, J. K. and Getis, A. 1995. Local spatial autocorrelation statistics: distributional issues and an application. Geographical Analysis, 27, 286–306. doi: 10.1111/j.1538-4632.1995.tb00912.x

See also

Examples

guerry_model <- guerry
guerry_lm <- lm(Crm_prs ~ Litercy, guerry_model)
guerry_model$predictions <- predict(guerry_lm, guerry_model)

ww_local_getis_ord_g(guerry_model, Crm_prs, predictions)
#> # A tibble: 85 × 3
#>    .metric           .estimator .estimate
#>    <chr>             <chr>          <dbl>
#>  1 local_getis_ord_g standard       0.913
#>  2 local_getis_ord_g standard       2.49 
#>  3 local_getis_ord_g standard       2.15 
#>  4 local_getis_ord_g standard      -1.58 
#>  5 local_getis_ord_g standard      -1.19 
#>  6 local_getis_ord_g standard      -1.68 
#>  7 local_getis_ord_g standard       0.627
#>  8 local_getis_ord_g standard      -1.60 
#>  9 local_getis_ord_g standard       0.964
#> 10 local_getis_ord_g standard      -2.71 
#> # … with 75 more rows
ww_local_getis_ord_g_pvalue(guerry_model, Crm_prs, predictions)
#> # A tibble: 85 × 3
#>    .metric                  .estimator .estimate
#>    <chr>                    <chr>          <dbl>
#>  1 local_getis_ord_g_pvalue standard     0.357  
#>  2 local_getis_ord_g_pvalue standard     0.0176 
#>  3 local_getis_ord_g_pvalue standard     0.0330 
#>  4 local_getis_ord_g_pvalue standard     0.112  
#>  5 local_getis_ord_g_pvalue standard     0.257  
#>  6 local_getis_ord_g_pvalue standard     0.0856 
#>  7 local_getis_ord_g_pvalue standard     0.562  
#>  8 local_getis_ord_g_pvalue standard     0.111  
#>  9 local_getis_ord_g_pvalue standard     0.304  
#> 10 local_getis_ord_g_pvalue standard     0.00691
#> # … with 75 more rows

wt <- ww_build_weights(guerry_model)

ww_local_getis_ord_g_vec(
  guerry_model$Crm_prs,
  guerry_model$predictions,
  wt = wt
)
#>  [1]  0.91288164  2.49305353  2.14692501 -1.57511235 -1.18988831 -1.67703329
#>  [7]  0.62706195 -1.60139345  0.96397088 -2.71475331 -3.05125861 -1.64429475
#> [13]  1.15584938 -2.90161984 -0.63376101  0.61005430  2.29888368 -0.28799789
#> [19]  1.53012357  1.22699107  0.59816133 -0.65331166  0.87501663 -1.75661589
#> [25]  0.39100454  0.52017062  1.16424138 -3.20781683 -2.29583912 -0.98116177
#> [31]  0.06685550 -2.13635236  2.30311768  0.63548278  1.99855789 -0.24366435
#> [37]  1.55058791  0.17231008  1.52062197 -0.22999799 -1.29501163  1.08298736
#> [43]  0.13063616 -1.21798455 -1.00549332 -2.29306942  0.89419782  1.19832028
#> [49]  2.03154442  0.87398122  2.88484032 -0.76194353  1.52453019  1.36764146
#> [55] -0.06618369  1.47010332  1.89277905  0.52529402  1.63007385  2.00852739
#> [61]  1.14763247 -0.91644996 -0.49669001 -1.85515687 -1.00553207 -0.86081555
#> [67]  1.63219209  1.04953518  2.51838763  1.71259713 -0.49967695  1.35387348
#> [73] -0.33336379 -0.07498653  0.56058846  1.54116260 -1.91772217 -1.92601432
#> [79] -1.62505600 -2.60384397  0.42262985  1.16021704  1.46229964  0.06056077
#> [85]  1.08651166
ww_local_getis_ord_g_pvalue_vec(
  guerry_model$Crm_prs,
  guerry_model$predictions,
  wt = wt
)
#>  [1] 0.3520454428 0.0169281576 0.0456194320 0.1213809468 0.2356873221
#>  [6] 0.1298839705 0.5134464784 0.0939851483 0.3540776416 0.0095529653
#> [11] 0.0074801388 0.0998363344 0.2396979838 0.0052251940 0.4916517907
#> [16] 0.5231543118 0.0261419040 0.8298269105 0.1334950545 0.2522464506
#> [21] 0.5703247487 0.5628307920 0.4039854865 0.0955473103 0.6771700952
#> [26] 0.6064596323 0.2312312170 0.0005308285 0.0350750779 0.3564773111
#> [31] 0.9365324730 0.0434975272 0.0256536539 0.5520962452 0.0517681958
#> [36] 0.7877747240 0.1496092316 0.9007078315 0.0928046348 0.8185749739
#> [41] 0.2115859588 0.2486524527 0.8994231204 0.2787058416 0.3066525627
#> [46] 0.0202432039 0.3806061055 0.2167217268 0.0533118481 0.3642734163
#> [51] 0.0047252088 0.4338439937 0.1196715896 0.1764061445 0.9093471255
#> [56] 0.1665499781 0.0452429286 0.6227874422 0.1030297749 0.0417938651
#> [61] 0.2910549103 0.4034412534 0.6012743473 0.0536110378 0.3279728815
#> [66] 0.4073282366 0.1178596545 0.3053693396 0.0117902732 0.1074315646
#> [71] 0.6443945571 0.1558252003 0.7203184068 0.9785927256 0.5807352300
#> [76] 0.1471011376 0.0399481674 0.0751198708 0.0998320185 0.0149890162
#> [81] 0.6780433950 0.2725022568 0.1468641113 0.9484784447 0.3145896955