Local Moran's I statistic

Calculate the local Moran's I statistic for model residuals. ww_local_moran_i() returns the statistic itself, while ww_local_moran_pvalue() returns the associated p value. ww_local_moran() returns both.

Usage

ww_local_moran_i(data, ...)

ww_local_moran_i_vec(
  truth,
  estimate,
  wt = NULL,
  alternative = "two.sided",
  na_rm = TRUE,
  ...
)

ww_local_moran_pvalue(data, ...)

ww_local_moran_pvalue_vec(
  truth,
  estimate,
  wt = NULL,
  alternative = "two.sided",
  na_rm = TRUE,
  ...
)

ww_local_moran(
  data,
  truth,
  estimate,
  wt = NULL,
  alternative = "two.sided",
  na_rm = TRUE,
  ...
)

Arguments

data

A data.frame containing the columns specified by the truth and estimate arguments.

...

Additional arguments passed to spdep::localmoran().

truth

The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate

The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

wt

A "listw" object, for instance as created with ww_build_weights().

alternative

a character string specifying the alternative hypothesis, must be one of greater, less or two.sided (default).

na_rm

A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A tibble with columns .metric, .estimator, and .estimate and nrow(data) rows of values. For grouped data frames, the number of rows returned will be the same as the number of groups. For _vec() functions, a numeric vector of length(truth) (or NA).

Examples

data(guerry, package = "sfdep")

guerry_modeled <- guerry
guerry_lm <- lm(crime_pers ~ literacy, guerry_modeled)
guerry_modeled$predictions <- predict(guerry_lm, guerry_modeled)

ww_local_moran_i(guerry_modeled, crime_pers, predictions)
#> # A tibble: 85 × 4
#>    .metric       .estimator .estimate                                   geometry
#>    <chr>         <chr>          <dbl>                             <MULTIPOLYGON>
#>  1 local_moran_i standard      0.530  (((931664 1785222, 930836 1785736, 929664…
#>  2 local_moran_i standard      0.858  (((931664 1785222, 930836 1785736, 929664…
#>  3 local_moran_i standard      0.759  (((931664 1785222, 930836 1785736, 929664…
#>  4 local_moran_i standard      0.732  (((931664 1785222, 930836 1785736, 929664…
#>  5 local_moran_i standard      0.207  (((931664 1785222, 930836 1785736, 929664…
#>  6 local_moran_i standard      0.860  (((931664 1785222, 930836 1785736, 929664…
#>  7 local_moran_i standard      0.692  (((931664 1785222, 930836 1785736, 929664…
#>  8 local_moran_i standard      1.69   (((931664 1785222, 930836 1785736, 929664…
#>  9 local_moran_i standard     -0.0109 (((931664 1785222, 930836 1785736, 929664…
#> 10 local_moran_i standard      0.710  (((931664 1785222, 930836 1785736, 929664…
#> # … with 75 more rows
ww_local_moran(guerry_modeled, crime_pers, predictions)
#> # A tibble: 170 × 4
#>    .metric       .estimator .estimate                                   geometry
#>    <chr>         <chr>          <dbl>                             <MULTIPOLYGON>
#>  1 local_moran_i standard      0.530  (((931664 1785222, 930836 1785736, 929664…
#>  2 local_moran_i standard      0.858  (((931664 1785222, 930836 1785736, 929664…
#>  3 local_moran_i standard      0.759  (((931664 1785222, 930836 1785736, 929664…
#>  4 local_moran_i standard      0.732  (((931664 1785222, 930836 1785736, 929664…
#>  5 local_moran_i standard      0.207  (((931664 1785222, 930836 1785736, 929664…
#>  6 local_moran_i standard      0.860  (((931664 1785222, 930836 1785736, 929664…
#>  7 local_moran_i standard      0.692  (((931664 1785222, 930836 1785736, 929664…
#>  8 local_moran_i standard      1.69   (((931664 1785222, 930836 1785736, 929664…
#>  9 local_moran_i standard     -0.0109 (((931664 1785222, 930836 1785736, 929664…
#> 10 local_moran_i standard      0.710  (((931664 1785222, 930836 1785736, 929664…
#> # … with 160 more rows