Evaluate metrics at multiple scales of aggregation

Evaluate metrics at multiple scales of aggregation

Usage

ww_multi_scale(
  data,
  truth,
  estimate,
  metrics = list(yardstick::rmse, yardstick::mae),
  grids = NULL,
  ...,
  na_action = na.fail,
  aggregation_function = mean,
  autoexpand_grid = TRUE
)

Arguments

data

A point geometry sf object containing the columns specified by the truth and estimate arguments.

truth

The column identifier for the true results (that is numeric). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For _vec() functions, a numeric vector.

estimate

The column identifier for the predicted results (that is also numeric). As with truth this can be specified different ways but the primary method is to use an unquoted variable name. For _vec() functions, a numeric vector.

metrics

Either a yardstick::metric_set() object, or a list of functions which will be used to construct a yardstick::metric_set() object specifying the performance metrics to evaluate at each scale.

grids

Optionally, a list of pre-computed sf or sfc objects specifying polygon boundaries to use for assessments.

...

Arguments passed to sf::st_make_grid(). You almost certainly should provide these arguments as lists. For instance, passing n = list(c(1, 2)) will create a single 1x2 grid; passing n = c(1, 2) will create a 1x1 grid and a 2x2 grid.

na_action

A function which indicates what should happen when the data contain NAs. The default is na.fail; you may wish to set it to na.omit or any of the functions from the "zoo" package. This function ignores the value of options("na.action") in order to make cross-computer (and cross-session) results more stable. This argument must be of length 1; you cannot pass multiple functions to na_action.

aggregation_function

The function to use to aggregate predictions and true values at various scales, by default mean(). You can pass any function which takes a single vector and returns a scalar.

autoexpand_grid

Boolean: if data is in geographic coordinates and grids aren't provided, the grids generated by sf::st_make_grid() may not contain all observations. If TRUE, this function will automatically expand generated grids by a tiny factor to attempt to capture all observations.

Value

A tibble with six columns: .metric, with the name of the metric that the row describes; .estimator, with the name of the estimator used, .estimate, with the output of the metric function; .grid_args, with the arguments passed to sf::st_make_grid() via ...

(if any), .grid, containing the grids used to aggregate predictions, and .notes, which will indicate any observations which were not used in a given assessment.

Details

The grid blocks can be controlled by passing arguments to sf::st_make_grid() via .... Some particularly useful arguments include:

cellsize: Target cellsize, expressed as the "diameter" (shortest straight-line distance between opposing sides; two times the apothem) of each block, in map units.

n: The number of grid blocks in the x and y direction (columns, rows).

square: A logical value indicating whether to create square (TRUE) or hexagonal (FALSE) cells.

If both cellsize and n are provided, then the number of blocks requested by n of sizes specified by cellsize will be returned, likely not lining up with the bounding box of data. If only cellsize is provided, this function will return as many blocks of size cellsize as fit inside the bounding box of data. If only n is provided, then cellsize will be automatically adjusted to create the requested number of cells.

This function can be used for geographic or projected coordinate reference systems and expect 2D data.

References

Riemann, R., Wilson, B. T., Lister, A., and Parks, S. (2010). "An effective assessment protocol for continuous geospatial datasets of forest characteristics using USFS Forest Inventory and Analysis (FIA) data." Remote Sensing of Environment 114(10), pp 2337-2352, doi: 10.1016/j.rse.2010.05.010 .

Examples

data(ames, package = "modeldata")
ames_sf <- sf::st_as_sf(ames, coords = c("Longitude", "Latitude"), crs = 4326)
ames_model <- lm(Sale_Price ~ Lot_Area, data = ames_sf)
ames_sf$predictions <- predict(ames_model, ames_sf)

ww_multi_scale(
  ames_sf,
  Sale_Price,
  predictions,
  n = list(
    c(10, 10),
    c(1, 1)
  ),
  square = FALSE
)
#> # A tibble: 4 × 6
#>   .metric .estimator .estimate .grid_args       .grid          .notes          
#>   <chr>   <chr>          <dbl> <list>           <list>         <list>          
#> 1 rmse    standard      82076. <tibble [1 × 2]> <sf [103 × 3]> <tibble [0 × 2]>
#> 2 mae     standard      60467. <tibble [1 × 2]> <sf [103 × 3]> <tibble [0 × 2]>
#> 3 rmse    standard      27862. <tibble [1 × 2]> <sf [5 × 3]>   <tibble [0 × 2]>
#> 4 mae     standard      23267. <tibble [1 × 2]> <sf [5 × 3]>   <tibble [0 × 2]>

# or, mostly equivalently
# (there will be a slight difference due to `autoexpand_grid = TRUE`)
grids <- list(
  sf::st_make_grid(ames_sf, n = c(10, 10), square = FALSE),
  sf::st_make_grid(ames_sf, n = c(1, 1), square = FALSE)
)
ww_multi_scale(ames_sf, Sale_Price, predictions, grids = grids)
#> # A tibble: 4 × 6
#>   .metric .estimator .estimate .grid_args       .grid          .notes          
#>   <chr>   <chr>          <dbl> <list>           <list>         <list>          
#> 1 rmse    standard      77853. <tibble [0 × 0]> <sf [103 × 3]> <tibble [0 × 2]>
#> 2 mae     standard      56796. <tibble [0 × 0]> <sf [103 × 3]> <tibble [0 × 2]>
#> 3 rmse    standard      27789. <tibble [0 × 0]> <sf [5 × 3]>   <tibble [0 × 2]>
#> 4 mae     standard      23173. <tibble [0 × 0]> <sf [5 × 3]>   <tibble [0 × 2]>